Remove css references to jekyll generated ids
Jekyll generates ids for headers when it builds the site. However, these ids are based on the text itself. If we reference those ids, and something changes the text of the header, it will break the css selection. After this change, the headers are selected based on their location in the dom, not their id.